+++ DQM/SiPixelPhase1TrackClusters/BuildFile.xml
@@ -1,5 +1,3 @@
 <use   name="DQM/SiPixelPhase1Common"/>
 <use   name="DataFormats/TrackerRecHit2D"/>
-<use   name="TrackingTools/TrackFitters"/>
-
 <flags   EDM_PLUGIN="1"/>
